backend/feedback/serializers.py
# app/serializers.py

from rest_framework import serializers
from .models import Feedback


class FeedbackSerializer(serializers.ModelSerializer):
    class Meta:
        model = Feedback
        fields = ['id', 'star_count', 'description', 'user', 'created_at']
