backend/requirements.txt
absl-py==2.1.0
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/altgraph-0.17.2-py2.py3-none-any.whl
anyio==3.6.2
asgiref==3.6.0
astunparse==1.6.3
audioread==3.0.1
cached-property==1.5.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
decorator==4.4.2
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dm-tree==0.1.8
drf-yasg==1.21.7
drf-yasg2==1.19.4
face-recognition-models==0.3.0
fastapi==0.89.1
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.47.2
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/future-0.18.2-py3-none-any.whl
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httptools==0.5.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
moviepy==1.0.3
msgpack==1.0.7
namex==0.0.7
numba==0.58.1
numpy==1.26.3
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.1
pandas==2.2.0
pillow==10.2.0
platformdirs==2.6.2
pooch==1.8.0
proglog==0.1.10
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.4
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.6.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0.1
qrcode==7.4
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rich==13.7.0
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.1
simplejson==3.19.1
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.1
sqlparse==0.4.3
starlette==0.22.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
watchfiles==0.18.1
websockets==10.4
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
